Split lateConstCheck out of cullOverReferences
