+++ compiler/resolution/Makefile.share
@@ -28,6 +28,7 @@ RESOLUTION_SRCS =                                              \
                   implementForallIntents.cpp                   \
                   implementForallIntents2.cpp                  \
                   initializerResolution.cpp                    \
+                  lateConstCheck.cpp                           \
                   lifetime.cpp                                 \
                   loopDetails.cpp                              \
                   lowerForalls.cpp                             \
